+++ CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [3.0.4]
+### Added
+- A centerA, centerB, and radius, constructor for Shape.Capsule2D.
+
 ### Fixed
 - Fixed inconsistent rendering due to change introduced in 3.0.2.
+- Fixed missing Arrow2D arrowheadScale application in one constructor.
 
 ## [3.0.3]
 ### Fixed